+++ src/Defer/DeferrableDirective.php
@@ -16,7 +16,6 @@ use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
 
 class DeferrableDirective extends BaseDirective implements DefinedDirective, FieldMiddleware
 {
-    const NAME = 'deferrable';
     const THE_DEFER_DIRECTIVE_CANNOT_BE_USED_ON_A_ROOT_MUTATION_FIELD = 'The @defer directive cannot be used on a root mutation field.';
     const THE_DEFER_DIRECTIVE_CANNOT_BE_USED_ON_A_NON_NULLABLE_FIELD = 'The @defer directive cannot be used on a Non-Nullable field.';
 
@@ -43,16 +42,6 @@ SDL;
     public function __construct(Defer $defer)
     {
         $this->defer = $defer;
-    }
-
-    /**
-     * Name of the directive.
-     *
-     * @return string
-     */
-    public function name(): string
-    {
-        return self::NAME;
     }
 
     /**