+++ src/Schema/Directives/InterfaceDirective.php
@@ -6,16 +6,6 @@ use Nuwave\Lighthouse\Support\Contracts\DefinedDirective;
 
 class InterfaceDirective extends BaseDirective implements DefinedDirective
 {
-    /**
-     * Name of the directive.
-     *
-     * @return string
-     */
-    public function name(): string
-    {
-        return 'interface';
-    }
-
     public static function definition(): string
     {
         return /* @lang GraphQL */ <<<'SDL'