+++ src/WhereConstraints/WhereConstraintsDirective.php
@@ -17,8 +17,6 @@ use Nuwave\Lighthouse\Support\Contracts\DefinedDirective;
 
 class WhereConstraintsDirective extends BaseDirective implements ArgBuilderDirective, ArgManipulator, DefinedDirective
 {
-    const NAME = 'whereConstraints';
-
     /**
      * @var \Nuwave\Lighthouse\WhereConstraints\Operator
      */
@@ -33,16 +31,6 @@ class WhereConstraintsDirective extends BaseDirective implements ArgBuilderDirec
     public function __construct(Operator $operator)
     {
         $this->operator = $operator;
-    }
-
-    /**
-     * Name of the directive.
-     *
-     * @return string
-     */
-    public function name(): string
-    {
-        return self::NAME;
     }
 
     public static function definition(): string